requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
albumentations>=1.3.0
opencv-python>=4.5.0
PyYAML>=5.4.1
scikit-learn>=0.24.2
matplotlib>=3.4.0
ultralytics  # for YOLO
detectron2  # for RCNN (install via conda or from source)
tensorboard>=2.6.0
tqdm>=4.62.0 